refactor: 일급 콜레션 사용.
