+++ src/main/java/bowling/domain/FinalFrame.java
@@ -1,14 +1,8 @@
 package bowling.domain;
 
-import bowling.domain.state.State;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
+import bowling.domain.state.Final;
 
 public class FinalFrame extends Frame {
-
-  private boolean canTriple = true;
-  private List<State> states = new ArrayList<>();
 
   public FinalFrame() {
     super(10);
@@ -16,52 +10,22 @@ public class FinalFrame extends Frame {
 
   @Override
   public FinalFrame roll(int pins) {
-    if (!states.isEmpty()) {
-      State last = states.get(states.size() - 1);
-      if (!last.isDone()) {
-        states.remove(states.size() - 1);
-        states.add(last.roll(pins));
+    if (state == null) {  // first
+      state = new Final();
+      state.roll(pins);
 
-        canTriple = false;
-        return this;
-      }
+      return this;
     }
 
-    states.add(State.of(pins));
+    state.roll(pins);
     return this;
-  }
-
-  @Override
-  public boolean isDone() {
-    if (this.states.size() == 3) {
-      return true;
-    }
-
-    if (impossibleTriple()) {
-      return true;
-    }
-
-    return false;
-  }
-
-  private boolean impossibleTriple() {
-    return this.states.size() == 2 && !canTriple;
-  }
-
-  @Override
-  public String symbol() {
-    return states.stream()
-        .map(State::symbol)
-        .collect(Collectors.joining(State.DELIMITER));
   }
 
   @Override
   public String toString() {
     return "FinalFrame{" +
-        "states=" + states +
-        ", number=" + number +
+        "number=" + number +
         ", state=" + state +
-        ", canTriple=" + canTriple +
         '}';
   }
 }
